Fix invisible text after using eraser
